requirements.txt
sentence-transformers
openai==0.28.0
faiss-cpu
